dataretrieval/requirements.txt
flask
confluent_kafka
datetime
mysql-connector-python
python-dotenv